Feature/unicode and 64bit support

+++ source/MiniViewsLib/MainDialog/GeneralTab.cpp
@@ -15,7 +15,7 @@ const std::string GeneralWindow::checkEditModeTooltip("(Ctrl Shift E)");
 const std::string GeneralWindow::addMiniViewTooltip("(Ctrl Shift A)");
 const std::string GeneralWindow::opacityTooltip("(Ctrl Shift -)/(Ctrl Shift +)");
 
-bool GeneralWindow::CreateThis(HWND hParent, u32 windowId)
+bool GeneralWindow::CreateThis(HWND hParent, u64 windowId)
 {
 	if ( getHandle() != NULL )
 		return SetParent(hParent);
@@ -70,7 +70,7 @@ void GeneralWindow::CreateSubWindows()
 	editOpacity.AddTooltip(opacityTooltip.c_str());
 	editOpacity.SetTextLimit(3);
 	editOpacity.SetEditNum(miniViews.GetOpacityLevel());
-	HWND hTransparencyBuddy= CreateWindowEx(NULL, UPDOWN_CLASS, NULL, WS_CHILDWINDOW | WS_VISIBLE | WS_DISABLED | UDS_SETBUDDYINT |
+	HWND hTransparencyBuddy = CreateWindowEx(NULL, UPDOWN_CLASS, NULL, WS_CHILDWINDOW | WS_VISIBLE | WS_DISABLED | UDS_SETBUDDYINT |
 		UDS_ALIGNRIGHT | UDS_ARROWKEYS | UDS_HOTTRACK, 0, 0, 0, 0, getHandle(), (HMENU)Id::EditOpacityBuddy, NULL, NULL);
 	SendMessage(hTransparencyBuddy, UDM_SETBUDDY, (WPARAM)editOpacity.getHandle(), NULL);
 	SendMessage(hTransparencyBuddy, UDM_SETRANGE32, 0, MAKELPARAM(255, 0));
